Create App.js
Hey guys not sure how to connect php storm to GitHub atm so I just pasted directly. 

This is just a template to what I think would look alright for the project lmk what you think to improve. 

I also posted the css file so just copy the two and see if it works.